+++ internal/provider/git/service.go
@@ -58,7 +58,7 @@ func resetService() {
 
 var errUntracked = errors.New("file has no git history")
 
-func (s *repoService) fileAge(relPath string) (int, error) {
+func (s *repoService) fileAge(relPath string) (int64, error) {
 	commits, err := s.fileCommitTimes(relPath)
 	if err != nil {
 		return 0, err
@@ -71,10 +71,10 @@ func (s *repoService) fileAge(relPath string) (int, error) {
 	oldest := commits[len(commits)-1]
 	age := time.Since(oldest)
 
-	return int(age.Seconds()), nil
+	return int64(age.Seconds()), nil
 }
 
-func (s *repoService) fileFreshness(relPath string) (int, error) {
+func (s *repoService) fileFreshness(relPath string) (int64, error) {
 	commits, err := s.fileCommitTimes(relPath)
 	if err != nil {
 		return 0, err
@@ -87,10 +87,10 @@ func (s *repoService) fileFreshness(relPath string) (int, error) {
 	newest := commits[0]
 	freshness := time.Since(newest)
 
-	return int(freshness.Seconds()), nil
+	return int64(freshness.Seconds()), nil
 }
 
-func (s *repoService) authorCount(relPath string) (int, error) {
+func (s *repoService) authorCount(relPath string) (int64, error) {
 	log, err := s.repo.Log(&gogit.LogOptions{FileName: &relPath})
 	if err != nil {
 		return 0, eris.Wrap(err, "failed to get git log")
@@ -112,7 +112,7 @@ func (s *repoService) authorCount(relPath string) (int, error) {
 		return 0, errUntracked
 	}
 
-	return len(authors), nil
+	return int64(len(authors)), nil
 }
 
 func (s *repoService) fileCommitTimes(relPath string) ([]time.Time, error) {
